src/requirements.txt
pyspark>=3.3.0,<3.5.0
pandas>=1.3.0,<2.1.0
openpyxl>=3.0.0
xlrd>=2.0.0
kafka-python
pymongo
streamlit
plotly
pandas